Created item drop rates map in `constants.py` + refactors
