+++ server/utils/nodes.py
@@ -1,0 +1,9 @@
+import random
+from constants import NODE_TYPE_WEIGHTS
+from models import NodeType
+
+
+def get_random_node_type() -> NodeType:
+    types = list(NODE_TYPE_WEIGHTS.keys())
+    weights = list(NODE_TYPE_WEIGHTS.values())
+    return random.choices(types, weights=weights, k=1)[0]
